requirements.txt
streamlit==1.34.0
scikit-learn==1.4.2
tensorflow==2.15.0
pillow==10.2.0
numpy==1.24.3
